Former-commit-id: bf051dbfad62b6895567358f9d61014ecf14a0c4

+++ Sources/Internal/UI/UIPackage.cpp
@@ -77,7 +77,7 @@ UIControl * UIPackage::GetControl(const String &name) const
 
 UIControl * UIPackage::GetControl(int32 index) const
 {
-    DVASSERT(0 <= index && index < controls.size());
+    DVASSERT(0 <= index && index < (int32)controls.size());
     return controls[index];
 }
